Slash commands

See merge request update-gitlab.yml/kotlinspringbot!60

+++ src/main/kotlin/uk/me/danielharman/kotlinspringbot/services/XkcdService.kt
@@ -34,7 +34,7 @@ class XkcdService(private val mongoOperations: MongoOperations) {
         return getComic()
     }
 
-    fun getComic(number: Int? = null): OperationResult<XkcdComic, String> {
+    fun getComic(number: Long? = null): OperationResult<XkcdComic, String> {
         val client = HttpClient(CIO)
         val url = if(number == null) latestUrl else comicUrl.format(number)
         val response = runBlocking { client.get<HttpResponse>(url) }
